Move train definition
Fix comments
